+++ molgenis-security/pom.xml
@@ -44,6 +44,11 @@
             <artifactId>molgenis-data-security</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.molgenis</groupId>
+            <artifactId>molgenis-web</artifactId>
+            <version>${project.version}</version>
+        </dependency>
         <!-- third party dependencies -->
         <dependency>
             <groupId>org.freemarker</groupId>
